made changes to the TextSuggestions side
