+++ lib/app/ui/streaming.dart
@@ -5,6 +5,7 @@ import '../widgets/bottom.message.box.dart';
 import '../widgets/side.stream.dart';
 import '../widgets/stars.dart';
 import '../widgets/stream.tile.dart';
+import '../widgets/text.suggestion.dart';
 
 class Streaming extends StatelessWidget {
   const Streaming({Key? key}) : super(key: key);
@@ -173,95 +174,4 @@ class Streaming extends StatelessWidget {
   }
 }
 
-class TextSuggestions extends StatelessWidget {
-  const TextSuggestions({
-    super.key,
-    required this.withAlpha,
-  });
-
-  final Color withAlpha;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Container(
-          width: 70,
-          height: 40,
-          decoration: BoxDecoration(
-            color: withAlpha,
-            borderRadius: const BorderRadius.all(Radius.circular(20)),
-          ),
-          child:  const Center(
-            child: Text(
-                'Hello!!',
-              style: TextStyle(
-                color: Colors.white,
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(
-          width: 10,
-        ),
-        Container(
-          width: 110,
-          height: 40,
-          decoration: BoxDecoration(
-            color: withAlpha,
-            borderRadius: const BorderRadius.all(Radius.circular(20)),
-          ),
-          child:  const Center(
-            child: Text(
-              'How\'s you',
-              style: TextStyle(
-                color: Colors.white,
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(
-          width: 05,
-        ),
-        CircleAvatar(
-          backgroundColor: withAlpha,
-          child: Image.asset(AppAssets.kEmojiUn),
-        ),
-        const SizedBox(
-          width: 05,
-        ),
-        CircleAvatar(
-          backgroundColor: withAlpha,
-          child: Image.asset(AppAssets.kEmojiD),
-        ),
-        const SizedBox(
-          width: 05,
-        ),
-        CircleAvatar(
-          backgroundColor: withAlpha,
-          child: Image.asset(AppAssets.kEmojiT),
-        ),
-        const SizedBox(
-          width: 05,
-        ),
-        Container(
-          width: 70,
-          height: 40,
-          decoration: BoxDecoration(
-            color: withAlpha,
-            borderRadius: const BorderRadius.all(Radius.circular(20)),
-          ),
-          child:  const Center(
-            child: Text(
-              'Great',
-              style: TextStyle(
-                color: Colors.white,
-              ),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
